refactored all types into one file
